Demo/NeoChat/requirements.txt
pyyaml
python-dotenv
requests